+++ inc/grain.hpp
@@ -12,5 +12,5 @@ struct Grain
     /* Growth tensor' coordinates */
     f_vec orientation;
 
-    double theta[27];
+    float theta[27];
 };
