+++ inc/setup.hpp
@@ -8,7 +8,7 @@ class Nucleator
     public:
     Nucleator() = default;
 
-    std::pair<double,double> nucleate(Domain& domain, _long_int y0, _int g0, const Configuration::Layer& layer);
+    std::pair<float,float> nucleate(Domain& domain, _long_int y0, _int g0, const Configuration::Layer& layer);
     std::unordered_map<uint16_t, Grain>& grains() {return m_grains;}
 
     private:
